Create Assignment 4d - Full work with strings array
In this program we create and work with array of strings(include work with pointer). We learn how to make a string array, how to add to it more lines, how to find something in it, how to twice\double alredy wroten line and how to print out the string array.
As well we learn how to dynemic allocate memory for arrays.
NOTE: Coded in VS2019.